applied google cpplint

+++ examples_oldgl_glfw/138_RigReadFileBiovision/main.cpp
@@ -5,59 +5,57 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#include "delfem2/rig_geo3.h"
 #include <iostream>
 #include <vector>
-#include <set>
-//
 #define GL_SILENCE_DEPRECATION
+#include <GLFW/glfw3.h>
+
+#include "delfem2/rig_geo3.h"
 #include "delfem2/glfw/viewer3.h"
 #include "delfem2/glfw/util.h"
 #include "delfem2/opengl/old/funcs.h"
 #include "delfem2/opengl/old/rigv3.h"
-#include <GLFW/glfw3.h>
 
 namespace dfm2 = delfem2;
 
 // -------------------------------------------
 
-int main(int argc,char* argv[])
-{
+int main(int argc, char *argv[]) {
   std::vector<dfm2::CRigBone> aBone;
   std::vector<dfm2::CChannel_BioVisionHierarchy> aChannelRotTransBone;
   int nframe = 0;
   std::vector<double> aValRotTransBone;
 
-  std::string path_bvh = std::string(PATH_INPUT_DIR)+"/jump.bvh";
+  std::string path_bvh = std::string(PATH_INPUT_DIR) + "/jump.bvh";
 //  std::cout << "path:" << path_bvh << std::endl;
   Read_BioVisionHierarchy(
-      aBone,aChannelRotTransBone,nframe,aValRotTransBone,
+      aBone, aChannelRotTransBone, nframe, aValRotTransBone,
       path_bvh);
   UpdateBoneRotTrans(aBone);
   std::cout << "nBone:" << aBone.size() << "   aCh:" << aChannelRotTransBone.size() << std::endl;
-  for(unsigned int ib=0;ib<aBone.size();++ib){
+  for (unsigned int ib = 0; ib < aBone.size(); ++ib) {
     std::cout << ib << " " << aBone[ib].name << std::endl;
   }
-  
+
   // -------------------------
-  
+
   delfem2::glfw::CViewer3 viewer;
   delfem2::glfw::InitGLOld();
   viewer.InitGL();
   viewer.camera.view_height = 15.0;
   viewer.camera.camera_rot_mode = delfem2::CCam3_OnAxisZplusLookOrigin<double>::CAMERA_ROT_MODE::TBALL;
   delfem2::opengl::setSomeLighting();
-  
+
   // -------------------------
-  
-  while(!glfwWindowShouldClose(viewer.window)){
+
+  while (!glfwWindowShouldClose(viewer.window)) {
     {
       static int iframe = 0;
-      const int nch = aChannelRotTransBone.size();
+      const size_t nch = aChannelRotTransBone.size();
       SetPose_BioVisionHierarchy(
           aBone, aChannelRotTransBone,
-          aValRotTransBone.data()+iframe*nch);
-      iframe = (iframe+1)%nframe;
+          aValRotTransBone.data() + iframe * nch);
+      iframe = (iframe + 1) % nframe;
     }
     // --------------------
     viewer.DrawBegin_oldGL();
